requirements.txt
streamlit
langchain-groq
python-dotenv
SpeechRecognition
soundfile
pygame
audio-recorder-streamlit
pyttsx3